requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
aiosqlite==0.20.0
python-dotenv==1.0.1
sentry-sdk==2.19.2
mcp[cli]>=1.0.0
Pillow>=10.0.0
